// src/app/core/models/partner.model.ts
export class Partner {
    id: number=0
    partnerTypeId:number=0;
    name: string='';
    uniqueIdentificationNumber: string='';
    vatNumber: string='';
    contactName: string='';
    country: string='';
    city: string='';
    address: string='';
    zipCode: string='';
    phone: string='';
    fax: string='';
    email: string='';
    website: string='';
    additionalInformation: string='';
    logo: string='';
}